+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Introduced auto generation of HPC work flows using `RAnEn::generateHPCWorkflow`.
 - Fixed a bug during subseting observation times in `RAnEn::plotAnalogTimeSeries`.
+- Add `operational` in `RAnEn::generateHPCWorkflow`
 
 # PAnEn 3.7.0